integrating new pyramid code
